Analyze code quality with Sonar during Travis CI build
Sonar analysis is only run when the branch belongs to the pitest/pitclipse
repository. Indeed, sending the results of the the analysis to SonarCloud
requires the encrypted SONAR token which is not provided by Travis CI when
the build comes from a fork.

The code checking whether the analysis can be run is located in .travis/build.sh
and is called by the .travis.yml configuration file.

I also struggled a bit to encrypt the token because encryption on travis-ci.com
is different from encryption on travis-ci.org. For the record here is what I did:
 > travis login --pro
 > travis encrypt --pro <token> -r pitest/pitclipse