Wait for pit to be done in the call of run
We relied on the console to contain a certain String to check wether pit
is done or not. Now we rely on the extension point which was made for
that and don't rely on the console view anymore. This should be alot
more stable, even if pit changes it output messages or the console view
is not open.

Also we waited sometimes several times for pit to finish in the same
test, example:
@Test
public void emptyClassAndEmptyTest() throws CoreException {
    removeMethods(FOO_CLASS, FOO_BAR_PACKAGE, TEST_PROJECT);
    removeMethods(FOO_TEST_CLASS, FOO_BAR_PACKAGE, TEST_PROJECT);
    runTest(FOO_TEST_CLASS, FOO_BAR_PACKAGE, TEST_PROJECT);
    consoleContains(0, 0, 100, 0, 0);
    mutationsAre(Collections.emptyList());
    coverageReportGenerated(0,100,100);
}

In this test we waited for pit in the methods consoleContains(...),
mutationsAre(...) and in coverageReportGenerated(...).